requirements.txt
PyYAML

# For pyright
autopep8
